w3c-reports/12092014/mac/runned-tests/websockets/unload-a-document/002-expected.txt

CONSOLE MESSAGE: line 18: TypeError: undefined is not an object (evaluating 'w.close')
FAIL: Timed out waiting for notifyDone to be called
Test requires popup blocker disabled
